Make ExternalAddressSelector pluggable
